Renamed Java classes to new org name
