+++ java/build.gradle.kts
@@ -18,15 +18,14 @@ java {
 
 tasks.named<Test>("test") {
     useJUnitPlatform()
-    systemProperty("java.library.path", "/Users/jlbabilino/Documents/TripleHelix/Programming/Repositories/HelixTrajectoryJ/helixtrajectorycpp/build/src/cpp")
 }
 
 publishing {
     publications {
         create<MavenPublication>("maven") {
-            groupId = "org.team2363"
-            artifactId = "helixtrajectory" + project.properties["platform_id"];
-            version = "0.0.0-pre5"
+            groupId = "org.sleipnirgroup"
+            artifactId = "trajoptlib" + project.properties["platform_id"];
+            version = "0.0.0-pre0"
 
             from(components["java"])
         }
@@ -34,8 +33,7 @@ publishing {
     repositories {
         maven {
             name = "github"
-            // id = "github"
-            url = uri("https://maven.pkg.github.com/jlbabilino/HelixTrajectory")
+            url = uri("https://maven.pkg.github.com/SleipnirGroup/TrajoptLib")
             credentials(PasswordCredentials::class)
         }
     }
