+++ java/src/main/java/org/sleipnirgroup/trajoptlib/Drivetrain.java
@@ -1,6 +1,6 @@
 // Copyright (c) TrajoptLib contributors
 
-package org.team2363.helixtrajectory;
+package org.sleipnirgroup.trajoptlib;
 
 import java.util.Objects;
