+++ java/src/main/java/org/sleipnirgroup/trajoptlib/HolonomicTrajectorySample.java
@@ -1,6 +1,6 @@
 // Copyright (c) TrajoptLib contributors
 
-package org.team2363.helixtrajectory;
+package org.sleipnirgroup.trajoptlib;
 
 public class HolonomicTrajectorySample extends TrajectorySample {
     public final double velocityX;
