+++ java/src/main/java/org/sleipnirgroup/trajoptlib/InvalidPathException.java
@@ -1,6 +1,6 @@
 // Copyright (c) TrajoptLib contributors
 
-package org.team2363.helixtrajectory;
+package org.sleipnirgroup.trajoptlib;
 
 public class InvalidPathException extends Exception {
     public InvalidPathException(String message) {
