+++ java/src/main/java/org/sleipnirgroup/trajoptlib/PluginLoadException.java
@@ -1,6 +1,6 @@
 // Copyright (c) TrajoptLib contributors
 
-package org.team2363.helixtrajectory;
+package org.sleipnirgroup.trajoptlib;
 
 public class PluginLoadException extends Exception {
     public PluginLoadException(String message, Throwable cause) {
