+++ java/src/main/java/org/sleipnirgroup/trajoptlib/SwerveModule.java
@@ -1,7 +1,6 @@
 // Copyright (c) TrajoptLib contributors
 
-package org.team2363.helixtrajectory;
-
+package org.sleipnirgroup.trajoptlib;
 
 public final class SwerveModule {
     public final double x;
