+++ java/src/main/java/org/sleipnirgroup/util/RuntimeLoader.java
@@ -3,7 +3,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package org.team2363.util;
+package org.sleipnirgroup.util;
 
 import java.io.File;
 import java.io.IOException;
